'修改由于使用通用类名引起的样式冲突，修改最大包邮金额不能为小于1的数'

+++ resource/wetpl/default/web/shopping/js/shoporder.js
@@ -533,7 +533,7 @@ checkdomain(domain,id);
                         //判断商品邮费
                         var postMoney = 0.00;
                         if(goods.postage_status == 2){
-                            if(goods.postage_max_money && parseFloat(goods.postage_max_money) !== 0){
+                            if(goods.postage_max_money && parseInt(goods.postage_max_money) !== 0){
                                 if(parseFloat(goods.price * goodsdata.goods_num) < parseFloat(goods.postage_max_money)){
                                     postMoney = parseFloat(goods.postage * goodsdata.goods_num)
                                 }
@@ -1030,7 +1030,7 @@ checkdomain(domain,id);
             var singlePrice = 0;
            price=price+parseFloat($(element).find("#one_price").find("span").text())*parseFloat($(element).find("#numm").val());
            if($(element).attr('postage_status') == 2){
-               if($(element).attr('postage_max_money') && parseFloat($(element).attr('postage_max_money')) !== 0 ){
+               if($(element).attr('postage_max_money') && parseInt($(element).attr('postage_max_money')) !== 0 ){
                    if(parseFloat($(element).find("#one_price").find("span").text() * $(element).find("#numm").val()) < parseFloat($(element).attr('postage_max_money'))){
                        singlePrice = parseFloat($(element).attr('postage') * $(element).find("#numm").val())
                    }
